+++ CHANGELOG.md
@@ -42,7 +42,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Converted `<manifold-data-resource-logo>` to use GraphQL (#635)
 - Converted `<manifold-resource-status>` to use GraphQL (#611)
 - Converted `<manifold-resource-list>` to use GraphQL (#627)
-
+- Converted `<manifold-data-sso-button>` to use GraphQL (#640)
 
 ### Breaking Changes:
 
@@ -97,6 +97,11 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Fixed success event after a resource rename to include potentially modified label (#547)
 
 ## [v0.5.12]
+
+### Breaking changes
+
+- `resource-id` renamed to `resource-label` in `<manifold-resource-card>`
+- `product-id` renamed to `product-label` in `<manifold-service-card>`
 
 ### Added
 
@@ -286,6 +291,13 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   component.
 - Improved `plan-selector` performance by reducing API calls for non-custom plans.
 
+## [v0.5.1]
+
+### Breaking changes
+
+- `resource-label` removed from `<manifold-data-product-logo>` in favor of new
+  `<manifold-data-resource-logo>` component
+
 ## [v0.5.0]
 
 ### Added
@@ -334,6 +346,13 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Changed how the `manifold-service-card` works to have it fetch the product unless given. This
   allows it to be used standalone or in the `marketplace`.
+
+### Breaking changes
+
+- `resource-name` renamed to `resource-label` in the following components:
+  `<manifold-resource-list>`, `<manifold-resource-plan>`, `<manifold-resource-product>`,
+  `<manifold-data-product-logo>`, `<manifold-data-product-name>`,
+  `<manifold-data-provision-button>`, `<manifold-data-resource-list>`
 
 ## [0.3.1]